Move `isActive()` from `HttpClientConnection` into `HttpConnection` - applicable to both server and client
